Updated Networking, Big Screen Font
Server & Client
Changed the networking to be based on the push pull paradigm so
that the client connects to the server.

Added line width to options that are found in server

Client
Added default plot to be sine wave.

Modified the dictionaries that are sent over to be
ordered dictionaries to preserve order when sending over